refactor: move npc ai behavior to npc module
